requirements.txt
# Telegram Bot Framework
aiogram==3.3.0
aiohttp==3.9.1
aiodns==3.1.1
Brotli==1.1.0

# OCR Libraries
easyocr==1.7.1
Pillow==10.1.0
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
scikit-image==0.22.0
scipy==1.11.4
imageio==2.33.1

# PyTorch (Required for EasyOCR)
torch==2.1.2
torchvision==0.16.2
pyyaml==6.0.1
ninja==1.11.1.1

# AI/LLM for data processing
google-generativeai==0.3.2
openai==1.6.1

# Database
SQLAlchemy==2.0.23
alembic==1.13.1
greenlet==3.0.3

# Data Processing
pandas==2.1.4
openpyxl==3.1.2
python-docx==1.1.0
et-xmlfile==1.1.0
lxml==5.0.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
regex==2023.12.25
pytz==2023.3.post1
tzdata==2023.4

# Logging
loguru==0.7.2

# Additional dependencies
numpy==1.26.2
requests==2.31.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0
six==1.16.0
magic-filter==1.0.12
pydantic==2.5.3
pydantic-core==2.14.6
typing-extensions==4.9.0
annotated-types==0.6.0
